Rename dashboardItems to dashboardItem
